requirements.txt
sv-ttk
pyserial
